+++ httpie/__init__.py
@@ -3,6 +3,6 @@ HTTPie: command-line HTTP client for the API era.
 
 """
 
-__version__ = '2.6.0.dev0'
+__version__ = '2.6.0'
 __author__ = 'Jakub Roztocil'
 __licence__ = 'BSD'